Add helper method for describing deployment to common package
