+++ common/strategy.go
@@ -1,0 +1,7 @@
+package common
+
+type Strategy interface {
+	Preconditions() []Precondition
+	Phases() []Phase
+	OnError(Deployment, error)
+}
